Observe ViewSets instead of QuerySets.
This enables us to properly handle native DRF filters together
with per-user serializers (like the new permissions-injecting
serializer in the new Resolwe).

Instead of specifying a QuerySet to observe, the new API takes
a special stripped HttpRequest object, which is picklable and
contains all the state that defines what is observed.