Fix deadlock when worker's ASGI_THREADS limit is 1
Observer evaluation is CPU-bound. Running multiple evaluations in
threads within the same process is not desired. When ASGI_THREADS
limit was set to 1, a deadlock would always occur because when the
worker emits results, async_to_sync(self.channel_layer.send) spawns
another thread to send the message. The deadlock may occur even
when several, but a fixed number of threads is allowed.

To avoid the deadlock we have refactored the WorkerConsumer to
an AsyncConsumer which runs the evaluation as synchronous code in
a separate thread using database_sync_to_async adapter.