Fix failing tests
when testing on Python 3.8. The issue was handling timeouts when
using async_timeout.timeout. As stated in its documentation
(https://pypi.org/project/async-timeout/), the asyncio.TimeoutError
is raised outside of context manager scope on timeout. However the
tests were expection CancelledError inside the context  manager scope
which is incorrect.